requirements.txt
google-api-python-client
requests
beautifulsoup4
prettyprint